+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 -->
 ## [Unreleased]
 ### Added
+
+- Methods in LoggingFolderCleanupTestBase to create a folder under the temp folder
 ### Fixed
 ### Changed
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.28.1637